Refactor: redis가 원천데이터 소스가 되도록 수정
- jpa를 사용하던 리포지토리 삭제
- Like테이블 삭제
- LikeRedisRepository.java -> LikeRedisRepository.java
