+++ src/main/java/com/stoury/domain/Like.java
@@ -1,25 +1,14 @@
 package com.stoury.domain;
 
-import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
-@Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "LIKES", uniqueConstraints = @UniqueConstraint(name = "LIKES_UNIQUE_CONSTRAINTS",
-        columnNames = {"MEMBER_ID", "FEED_ID"}))
 public class Like {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(optional = false)
     private Member member;
 
-    @ManyToOne(optional = false)
     private Feed feed;
 
     public Like(Member member, Feed feed) {
